make find_all_shaders.py include default shaders
Maps require some shaders that don't fit the normal
naming scheme and are not in the shader database.
This change add special handling for them by first
checking if they can be found in the packages that
are extraced anyway and only if that doesn't work
we extract them from a packages that is known to contain
them.