port_ops_file_cs1: log errors to a test file and don't abort on first failure
With this change errors are now logged to a file (in addition to the
console) and the user is informed at the end if there has been an error.
The script also no longer aborts on the first failure but continues
after logging the error. There are some things that are not supported by
the underlying (un)pack tools and that need manual fixes. To make it
possible to at least port the rest the script now skips over such
errors.